requirements.txt
# Core dependencies for device-orchestra
opencv-python>=4.8.0
pyserial>=3.5
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pytest>=7.4.0
psutil>=5.9.0
click>=8.1.0

# Async dependencies for new architecture
aiohttp>=3.9.0
serial-asyncio>=0.6
pytest-asyncio>=0.21.0

# Development dependencies
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional dependencies for specific platforms
# python-periphery>=2.4.1  # For OrangePi/RPi GPIO
# paramiko>=3.3.0  # For SSH adapter
# requests>=2.31.0  # For REST API adapter
